Refactor : VulDB Skeleton 컴포넌트 모아놓은 파일 생성
