+++ src/components/vulnerability-db/ArticleDetail.tsx
@@ -4,17 +4,12 @@ import { useVulDBPostsStore } from "@/stores/useVulDBPostsStore";
 import { VulDBPost } from "@/types/post";
 import { isCertCCContentType, isCnnvdContentType } from "@/types/typeGuards";
 import Link from "next/link";
-import {
-  ArticleDetailHeader,
-  SkeletonDetailHeader,
-} from "./ArticleDetailHeader";
+import { ArticleDetailHeader } from "./ArticleDetailHeader";
 import { useLayoutEffect, useState } from "react";
-
-function SkeletonContent() {
-  return (
-    <div className="h-[62.5rem] w-[80.25rem] animate-pulse rounded bg-gray-200"></div>
-  );
-}
+import {
+  ArticleDetailContentSkeleton,
+  ArticleDetailHeaderSkeleton,
+} from "./VulDBSkeleton";
 
 type SectionProps = {
   title: string;
@@ -127,8 +122,8 @@ export default function ArticleDetail({
   if (!isMounted) {
     return (
       <>
-        <SkeletonDetailHeader />
-        <SkeletonContent />
+        <ArticleDetailHeaderSkeleton />
+        <ArticleDetailContentSkeleton />
       </>
     );
   }
