+++ src/components/vulnerability-db/ArticleDetailHeader.tsx
@@ -7,37 +7,7 @@ import { useLayoutEffect, useMemo, useState } from "react";
 import { Label } from "../ui/Label";
 import VulDBPin from "./VulDBPin";
 import VulDBShare from "./VulDBShare";
-
-export function LabelSkeleton() {
-  return (
-    <div className="inline-flex h-[2.188rem] w-16 animate-pulse items-center justify-center rounded-full bg-gray-200 px-3 py-2 leading-[1.216rem]"></div>
-  );
-}
-export function SkeletonDetailHeader() {
-  return (
-    <section className="w-full max-w-[82.125rem] border-b border-b-line-default p-[1.75rem_0_3.75rem_0] px-[1rem]">
-      <LabelSkeleton />
-      <div className="mb-8 mt-[1.261rem] h-9 w-4/5 rounded bg-gray-200"></div>
-      <div className="flex items-center justify-between text-xl font-normal leading-[1.512rem] text-gray-default">
-        <div className="flex gap-9 tracking-[-0.01em]">
-          <div className="h-6 w-40 rounded bg-gray-200"></div>
-          <div className="flex gap-1">
-            <div className="h-6 w-24 rounded bg-gray-200"></div> |
-            <div className="h-6 w-40 rounded bg-gray-200"></div>
-          </div>
-          <div className="flex gap-1">
-            <div className="h-6 w-24 rounded bg-gray-200"></div> |
-            <div className="h-6 w-40 rounded bg-gray-200"></div>
-          </div>
-        </div>
-        <div className="relative flex gap-[1.625rem]">
-          <div className="h-7 w-7 rounded-full bg-gray-200"></div>
-          <div className="h-7 w-7 rounded-full bg-gray-200"></div>
-        </div>
-      </div>
-    </section>
-  );
-}
+import { LabelSkeleton } from "./VulDBSkeleton";
 
 export function ArticleDetailHeader({
   post,
