+++ src/components/vulnerability-db/SimilarInfoPosts.tsx
@@ -14,33 +14,9 @@ import { VulDBPinnedInfo, VulDBPostWithChip } from "@/types/post";
 import { isCertCCContentType, isCnnvdContentType } from "@/types/typeGuards";
 import Link from "next/link";
 import { useLayoutEffect, useMemo, useState } from "react";
-import { LabelSkeleton } from "./ArticleDetailHeader";
 import VulDBPin from "./VulDBPin";
 import VulDBShare from "./VulDBShare";
-
-function SkeletonCard() {
-  return (
-    <Card
-      variant="article"
-      size="short"
-      className="h-[17.375rem] animate-pulse rounded-[1.25rem]"
-    >
-      <div className="flex-col-center-start relative h-fit flex-wrap items-center gap-2">
-        <LabelSkeleton />
-        <div className="mb-2 mt-2 h-7 w-full rounded bg-gray-200" />
-        <div className="h-7 w-full rounded bg-gray-200" />
-      </div>
-      <div className="h-6 w-full rounded bg-gray-200" />
-      <div className="flex items-center justify-between">
-        <div className="inline-flex gap-4">
-          <div className="h-8 w-8 rounded-full bg-gray-200" />
-          <div className="h-8 w-8 rounded-full bg-gray-200" />
-        </div>
-        <div className="h-5 w-1/6 rounded bg-gray-200" />
-      </div>
-    </Card>
-  );
-}
+import { SimilarInfoCardSkeleton } from "./VulDBSkeleton";
 
 function SimilarInfoPostCard({
   post,
@@ -147,7 +123,7 @@ export default function SimilarInfoPosts({
         <ul className="grid grid-cols-3 gap-9">
           {Array.from({ length: 6 }).map((_, index) => (
             <li key={index}>
-              <SkeletonCard />
+              <SimilarInfoCardSkeleton />
             </li>
           ))}
         </ul>
