+++ src/components/vulnerability-db/VulDBDashboard.tsx
@@ -11,7 +11,8 @@ import { useEffect, useState, useTransition } from "react";
 import Button from "../ui/Button";
 import { Modal, ModalTitle, ModalTitleWrapper } from "../ui/Modal";
 import Pagination from "../ui/Pagination";
-import VulDBList, { VulDBListCardSkeleton } from "./VulDBList";
+import VulDBList from "./VulDBList";
+import { VulDBListCardSkeleton } from "./VulDBSkeleton";
 
 // Hot과 New chip에 따른 필터링 적용
 const filterByChip = (
