+++ src/components/vulnerability-db/VulDBList.tsx
@@ -12,7 +12,6 @@ import { formatTimestampAsDaysAgo } from "@/lib/utils";
 import { VulDBPinnedInfo, VulDBPostWithChip } from "@/types/post";
 import { isCertCCContentType, isCnnvdContentType } from "@/types/typeGuards";
 import Link from "next/link";
-import { LabelSkeleton } from "./ArticleDetailHeader";
 import VulDBPin from "./VulDBPin";
 import VulDBShare from "./VulDBShare";
 
@@ -32,35 +31,6 @@ export function NoPostsMessage({
         <span className="text-[1.5rem] text-gray-default">{solution}</span>
       </p>
     </div>
-  );
-}
-
-export function VulDBListCardSkeleton() {
-  return (
-    <Card
-      variant="article"
-      size="long"
-      className="h-[16.875rem] animate-pulse rounded-[1.25rem] border-line-light"
-    >
-      <CardHeader className="gap-0">
-        <div className="mb-3 flex items-center gap-2">
-          <LabelSkeleton />
-          {/* Label */}
-          <div className="h-7 w-4/5 rounded bg-gray-200"></div> {/* Title */}
-        </div>
-        <div className="h-5 w-1/12 rounded bg-gray-200"></div> {/* Subtitle */}
-      </CardHeader>
-      <CardContent className="bg-gray-200" />
-      <CardFooter>
-        <div className="flex gap-3">
-          <div className="h-7 w-7 rounded-full bg-gray-200"></div>
-          {/* IconPin */}
-          <div className="h-7 w-7 rounded-full bg-gray-200"></div>
-          {/* IconExternalLink */}
-        </div>
-        <div className="h-6 w-16 rounded bg-gray-200"></div> {/* Date */}
-      </CardFooter>
-    </Card>
   );
 }
