+++ src/components/vulnerability-db/VulDBSkeleton.tsx
@@ -1,0 +1,97 @@
+import {
+  Card,
+  CardContent,
+  CardFooter,
+  CardHeader,
+} from "@/components/ui/Card";
+
+export function LabelSkeleton() {
+  return (
+    <div className="inline-flex h-[2.188rem] w-16 animate-pulse items-center justify-center rounded-full bg-gray-200 px-3 py-2 leading-[1.216rem]"></div>
+  );
+}
+
+export function VulDBListCardSkeleton() {
+  return (
+    <Card
+      variant="article"
+      size="long"
+      className="h-[16.875rem] animate-pulse rounded-[1.25rem] border-line-light"
+    >
+      <CardHeader className="gap-0">
+        <div className="mb-3 flex items-center gap-2">
+          <LabelSkeleton />
+          {/* Label */}
+          <div className="h-7 w-4/5 rounded bg-gray-200"></div> {/* Title */}
+        </div>
+        <div className="h-5 w-1/12 rounded bg-gray-200"></div> {/* Subtitle */}
+      </CardHeader>
+      <CardContent className="bg-gray-200" />
+      <CardFooter>
+        <div className="flex gap-3">
+          <div className="h-7 w-7 rounded-full bg-gray-200"></div>
+          {/* IconPin */}
+          <div className="h-7 w-7 rounded-full bg-gray-200"></div>
+          {/* IconExternalLink */}
+        </div>
+        <div className="h-6 w-16 rounded bg-gray-200"></div> {/* Date */}
+      </CardFooter>
+    </Card>
+  );
+}
+
+export function ArticleDetailHeaderSkeleton() {
+  return (
+    <section className="w-full max-w-[82.125rem] border-b border-b-line-default p-[1.75rem_0_3.75rem_0] px-[1rem]">
+      <LabelSkeleton />
+      <div className="mb-8 mt-[1.261rem] h-9 w-4/5 rounded bg-gray-200"></div>
+      <div className="flex items-center justify-between text-xl font-normal leading-[1.512rem] text-gray-default">
+        <div className="flex gap-9 tracking-[-0.01em]">
+          <div className="h-6 w-40 rounded bg-gray-200"></div>
+          <div className="flex gap-1">
+            <div className="h-6 w-24 rounded bg-gray-200"></div> |
+            <div className="h-6 w-40 rounded bg-gray-200"></div>
+          </div>
+          <div className="flex gap-1">
+            <div className="h-6 w-24 rounded bg-gray-200"></div> |
+            <div className="h-6 w-40 rounded bg-gray-200"></div>
+          </div>
+        </div>
+        <div className="relative flex gap-[1.625rem]">
+          <div className="h-7 w-7 rounded-full bg-gray-200"></div>
+          <div className="h-7 w-7 rounded-full bg-gray-200"></div>
+        </div>
+      </div>
+    </section>
+  );
+}
+
+export function ArticleDetailContentSkeleton() {
+  return (
+    <div className="h-[62.5rem] w-[80.25rem] animate-pulse rounded bg-gray-200"></div>
+  );
+}
+
+export function SimilarInfoCardSkeleton() {
+  return (
+    <Card
+      variant="article"
+      size="short"
+      className="h-[17.375rem] animate-pulse rounded-[1.25rem]"
+    >
+      <div className="flex-col-center-start relative h-fit flex-wrap items-center gap-2">
+        <LabelSkeleton />
+        <div className="mb-2 mt-2 h-7 w-full rounded bg-gray-200" />
+        <div className="h-7 w-full rounded bg-gray-200" />
+      </div>
+      <div className="h-6 w-full rounded bg-gray-200" />
+      <div className="flex items-center justify-between">
+        <div className="inline-flex gap-4">
+          <div className="h-8 w-8 rounded-full bg-gray-200" />
+          <div className="h-8 w-8 rounded-full bg-gray-200" />
+        </div>
+        <div className="h-5 w-1/6 rounded bg-gray-200" />
+      </div>
+    </Card>
+  );
+}
